+++ env.go
@@ -1,12 +1,11 @@
 package flotilla
 
 import (
-	"bufio"
-	"bytes"
 	"fmt"
 	"os"
 	"path/filepath"
-	"reflect"
+
+	"lcl/flotilla/session"
 
 	flag "gopkg.in/alecthomas/kingpin.v1"
 )
@@ -28,77 +27,45 @@ var (
 type (
 	// The engine environment
 	Env struct {
-		staticdirectories []string
-		ctxfunctions      map[string]interface{}
-		Mode              int
-		SessionManager    *Manager
-		Conf
+		Mode int
+		Store
+		SessionManager *session.Manager
 		Assets
 		Templator
+		flotilla     map[string]Flotilla
+		ctxfunctions map[string]interface{}
 	}
 )
 
 func BaseEnv() *Env {
-	e := &Env{Conf: make(map[string]string)}
+	e := &Env{Store: make(Store)}
 	e.Templator = NewTemplator(e)
+	e.ctxfunctions = make(map[string]interface{})
 	e.AddCtxFuncs(builtinctxfuncs)
+	e.Store.adddefault("secret", "key", "defaultsecretkeypleasechangethis")
 	return e
 }
 
 // Merges an env instance with the calling env
-func (env *Env) MergeEnv(mergeenv *Env) {
-	env.LoadConfMap(mergeenv.Conf)
-	for _, fs := range mergeenv.Assets {
+func (env *Env) MergeEnv(me *Env) {
+	env.MergeStore(me.Store)
+	for _, fs := range me.Assets {
 		env.Assets = append(env.Assets, fs)
 	}
-	for _, dir := range mergeenv.StaticDirs() {
+	for _, dir := range me.StaticDirs() {
 		env.AddStaticDir(dir)
 	}
-	env.AddTemplatesDir(mergeenv.Templator.ListTemplateDirs()...)
-	env.AddCtxFuncs(mergeenv.ctxfunctions)
+	env.AddTemplatesDir(me.Templator.ListTemplateDirs()...)
+	env.AddCtxFuncs(me.ctxfunctions)
 }
 
-// Loads a conf file into the env from the engine
-func (engine *Engine) EnvConfFile(flpth string) bool {
-	err := engine.Env.LoadConfFile(flpth)
-	if err == nil {
-		return true
+// Merges an Store instance with the Env's Store
+func (env *Env) MergeStore(s Store) {
+	for k, v := range s {
+		if _, ok := env.Store[k]; !ok {
+			env.Store[k] = v
+		}
 	}
-	return false
-}
-
-// Loads a conf map into the env from engine
-func (engine *Engine) EnvConfMap(m map[string]string) bool {
-	err := engine.Env.LoadConfMap(m)
-	if err == nil {
-		return true
-	}
-	return false
-}
-
-// Loads a conf file into the env
-func (env *Env) LoadConfFile(filename string) (err error) {
-	file, err := os.Open(filename)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-	reader := bufio.NewReader(file)
-	err = env.Conf.parse(reader, filename)
-	return err
-}
-
-// Loads a conf as byte into the env
-func (env *Env) LoadConfByte(b []byte, name string) (err error) {
-	reader := bufio.NewReader(bytes.NewReader(b))
-	err = env.Conf.parse(reader, name)
-	return err
-}
-
-// Loads a conf as map into the env
-func (env *Env) LoadConfMap(m map[string]string) (err error) {
-	err = env.Conf.parsemap(m)
-	return err
 }
 
 // Sets the running mode for the engine env by a string
@@ -115,23 +82,25 @@ func (env *Env) SetMode(value string) {
 	}
 }
 
-// All static directories specified for the engine including those set directly
-// with the engine and those set in a configuration file
+// A string array of static dirs set in env.Store["staticdirectories"]
 func (env *Env) StaticDirs() []string {
-	dirs := env.staticdirectories
-	if c, err := env.List("staticdirectories"); err == nil {
-		for _, d := range c {
-			dirs = append(dirs, d)
+	if static, ok := env.Store["staticdirectories"]; ok {
+		if ret, err := static.getList(); err == nil {
+			return ret
 		}
 	}
-	return dirs
+	return []string{}
 }
 
 // Adds a static directory to be searched when a static route is accessed.
-func (env *Env) AddStaticDir(dir string) {
-	env.staticdirectories = dirAdd(dir, env.staticdirectories)
+func (env *Env) AddStaticDir(dirs ...string) {
+	if _, ok := env.Store["staticdirectories"]; !ok {
+		env.Store.add("", "staticdirectories", "")
+	}
+	env.Store["staticdirectories"].updateList(dirs...)
 }
 
+// Listing of templator template directories
 func (env *Env) TemplateDirs() []string {
 	return env.Templator.ListTemplateDirs()
 }
@@ -141,29 +110,11 @@ func (env *Env) AddTemplatesDir(dirs ...string) {
 	env.Templator.UpdateTemplateDirs(dirs...)
 }
 
-// Adds cross-handler functions from a map
+// Adds cross-handler functions
 func (env *Env) AddCtxFuncs(fns map[string]interface{}) {
 	for k, v := range fns {
-		env.AddCtxFunc(k, v)
+		env.ctxfunctions[k] = v
 	}
-}
-
-// Adds a cross-handler function by name/interface
-func (env *Env) AddCtxFunc(name string, fn interface{}) {
-	if env.ctxfunctions == nil {
-		env.ctxfunctions = make(map[string]interface{})
-	}
-
-	env.ctxfunctions[name] = fn
-}
-
-// All env ctxfunctions available as reflect.Value(for use by *Ctx)
-func (env *Env) CtxFunctions() map[string]reflect.Value {
-	m := make(map[string]reflect.Value)
-	for k, v := range env.ctxfunctions {
-		m[k] = valueFunc(v)
-	}
-	return m
 }
 
 func (env *Env) parseFlags() {
@@ -172,20 +123,13 @@ func (env *Env) parseFlags() {
 	env.SetMode(*flagMode)
 }
 
-func (env *Env) Secret() (secret string) {
-	secret = env.Conf["secret"]
-	if len(secret) == 0 {
-		secret = "secretpleasechangethis"
-	}
-	return secret
-}
-
 func (env *Env) defaultsessionconfig() string {
-	return fmt.Sprintf(`{"cookieName":"flotillasessionid","enableSetCookie":false,"gclifetime":3600,"ProviderConfig":"{\"cookieName\":\"flotillasessionid\",\"securityKey\":\"%s\"}"}`, env.Secret())
+	secret := env.Store["secret_key"].value
+	return fmt.Sprintf(`{"cookieName":"flotillasessionid","enableSetCookie":false,"gclifetime":3600,"ProviderConfig":"{\"cookieName\":\"flotillasessionid\",\"securityKey\":\"%s\"}"}`, secret)
 }
 
-func (env *Env) defaultsessionmanager() (*Manager, error) {
-	return NewManager("cookie", env.defaultsessionconfig())
+func (env *Env) defaultsessionmanager() (*session.Manager, error) {
+	return session.NewManager("cookie", env.defaultsessionconfig())
 }
 
 func (env *Env) SessionInit() {
@@ -194,7 +138,7 @@ func (env *Env) SessionInit() {
 		if err == nil {
 			env.SessionManager = sm
 		} else {
-			panic(fmt.Sprintf("Problem with default sesson manager: %s", err))
+			panic(fmt.Sprintf("Problem with default session manager: %s", err))
 		}
 	}
 	go env.SessionManager.GC()