+++ CHANGELOG.md
@@ -1,20 +1,27 @@
 ##Changelog
 
-###Flotilla 0.0.2 (0.0.0)
+###Flotilla 0.0.3 (tbd) 
+
+
+###Flotilla 0.0.2 (9.24.2014)
 
 - extend Ctx with cross handler functions
 - simple flag parsing for run mode (production, development, testing)
 - cookie based sessions as a default with capacity for adding different backends
+- folded router & some lower-level-but-not-in-net/http(such as httpexceptions)
+  functions into another package: [Engine](https://github.com/thrisp/engine)
+- url formatting by route, i.e. creating urls by route name & parameters
+
 
 ###Flotilla 0.0.1 (8.20.2014)
 
 - reforked, renamed as Flotilla
 - ini style configuration read into app environment
-- basic Flotilla interface for extension
+- basic Flotilla interface for extension of routes & env
 - basic Jingo templating
 - provisions for binary static or template Asset inclusion per engine
  
 
-###Fleet 0.0 (7.22.2014)
+###Fleet 0.0.0 (7.22.2014)
 
 - forked from https://github.com/gin-gonic/gin